feat: added blog comments section to single blog post

+++ Explorer/src/app/feature-modules/blog/blog-comment/blog-comment.component.ts
@@ -1,7 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
 import { BlogComment } from '../model/blogComment.model';
 import { BlogService } from '../blog.service';
 import { PagedResults } from 'src/app/shared/model/paged-results.model';
+import { AuthService } from 'src/app/infrastructure/auth/auth.service';
 
 @Component({
   selector: 'xp-blog-comment',
@@ -11,15 +12,18 @@ import { PagedResults } from 'src/app/shared/model/paged-results.model';
 
 export class BlogCommentComponent implements OnInit {
 
+  @Input() blogPostId: number | undefined;
   blogComments: BlogComment[] = [];
   selectedBlogComment: BlogComment;
   shouldRenderBlogCommentForm: boolean = false;
   shouldEdit: boolean = false;
   showOptions: number | null = null;
+  userId: number;
 
-  constructor(private service: BlogService) { }
+  constructor(private service: BlogService, private authService: AuthService) { }
 
   ngOnInit(): void {
+    this.userId = this.authService.user$.value.id;
     this.getBlogComments();
   }
 
@@ -29,7 +33,6 @@ export class BlogCommentComponent implements OnInit {
         this.blogComments = result.results;
       },
       error: () => {
-        // handle error
       }
     });
   }
@@ -46,17 +49,25 @@ export class BlogCommentComponent implements OnInit {
     return this.showOptions === index;
   }
 
-  editComment(blogComment: BlogComment): void {
+  onEditClicked(blogComment: BlogComment): void {
     this.selectedBlogComment = blogComment;
     this.shouldRenderBlogCommentForm = true;
     this.shouldEdit = true;
   }
 
   deleteComment(id: number): void {
-    this.service.deleteBlogComment(id).subscribe({
-      next: () => {
-        this.getBlogComments();
-      },
-    })
+    const result = window.confirm('Are you sure you want to delete your comment?');
+    if(result) {
+      this.service.deleteBlogComment(id).subscribe({
+        next: () => {
+          this.getBlogComments();
+          this.onEditingFinished(false);
+        },
+      })
+    }
+  }
+
+  onEditingFinished(value: boolean): void {
+    this.shouldEdit = value;
   }
 }
